Refactor taxi template to use absolute import.
PiperOrigin-RevId: 300742532
