+++ RELEASE.md
@@ -3,6 +3,7 @@
 ## Major Features and Improvements
 
 ## Bug fixes and other changes
+*   Replaced relative import with absolute import in generated templates.
 
 ### Deprecations
