+++ tfx/experimental/templates/taxi/models/preprocessing_test.py
@@ -19,23 +19,14 @@ from __future__ import print_function
 
 import tensorflow as tf
 
-from tfx.experimental.templates.taxi import features
+from tfx.experimental.templates.taxi.models import preprocessing
 
 
-class FeaturesTest(tf.test.TestCase):
+class PreprocessingTest(tf.test.TestCase):
 
-  def testNumberOfBucketFeatureBucketCount(self):
-    self.assertEqual(
-        len(features.BUCKET_FEATURE_KEYS),
-        len(features.BUCKET_FEATURE_BUCKET_COUNT))
-    self.assertEqual(
-        len(features.CATEGORICAL_FEATURE_KEYS),
-        len(features.CATEGORICAL_FEATURE_MAX_VALUES))
-
-  def testTransformedNames(self):
-    names = ["f1", "cf"]
-    self.assertEqual(["f1_xf", "cf_xf"], features.transformed_names(names))
+  def testPreprocessingFn(self):
+    self.assertTrue(callable(preprocessing.preprocessing_fn))
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
   tf.test.main()
