+++ tfx/experimental/templates/taxi/pipeline/configs.py
@@ -19,8 +19,6 @@ This file defines environments for a TFX taxi pipeline.
 
 import os  # pylint: disable=unused-import
 
-from tfx.proto import trainer_pb2
-
 # TODO(b/149347293): Move more TFX CLI flags into python configuration.
 
 # Pipeline name will be used to identify this pipeline.
@@ -40,11 +38,11 @@ GCS_BUCKET_NAME = 'YOUR_GCS_BUCKET_NAME'
 # GCP_PROJECT_ID = 'YOUR_GCP_PROJECT_ID'
 # GCP_REGION = 'YOUR_GCP_REGION'  # ex) 'us-central1'
 
-PREPROCESSING_FN = 'preprocessing.preprocessing_fn'
-TRAINER_FN = 'model.trainer_fn'
+PREPROCESSING_FN = 'models.preprocessing.preprocessing_fn'
+TRAINER_FN = 'models.estimator.model.trainer_fn'
 
-TRAIN_ARGS = trainer_pb2.TrainArgs(num_steps=100)
-EVAL_ARGS = trainer_pb2.EvalArgs(num_steps=100)
+TRAIN_NUM_STEPS = 100
+EVAL_NUM_STEPS = 100
 
 # Beam args to use BigQueryExampleGen.
 # TODO(step 7): (Optional) Uncomment here to provide GCP related configs for
@@ -89,8 +87,8 @@ _query_sample_rate = 0.0001  # Generate a 0.01% random sample.
 #    query_sample_rate=_query_sample_rate)
 
 # Beam args to run data processing on DataflowRunner.
-# TODO(b/151114974) Remove `disk_size_gb` flag after default is increased.
-# TODO(b/151116587) Remove `shuffle_mode` flag after default is changed.
+# TODO(b/151114974): Remove `disk_size_gb` flag after default is increased.
+# TODO(b/151116587): Remove `shuffle_mode` flag after default is changed.
 # TODO(step 8): (Optional) Uncomment below to use Dataflow.
 # BEAM_PIPELINE_ARGS = [
 #    '--project=' + GCP_PROJECT_ID,
